Revert "Feature: User Authentication (#1)"
This reverts commit db2cbd61010414fbd3976898ba1e5df9b915fe39.
