+++ .gitignore
@@ -3,11 +3,6 @@ target
 
 # Ignore vscode directory
 .vscode
-.VSCodeCounter
 
 # Ignore compiled static frontend
-static/frontend/
-static/styles/app.css
-
-# Ignore temp db
-db.db
+static/frontend
