+++ backend/src/args.rs
@@ -7,20 +7,12 @@ use clap::Parser;
 #[command()]
 pub struct ProgramArgs {
     /// The port to serve on
-    #[arg(long)]
+    #[arg(short, long)]
     pub port: Option<u16>,
 
     /// The directory to serve static files from
-    #[arg(long)]
+    #[arg(short, long)]
     pub static_dir: Option<PathBuf>,
-
-    /// The url to the database
-    #[arg(long)]
-    pub database_url: Option<String>,
-
-    /// Create a super user with a given name and password separated by a colon
-    #[arg(long)]
-    pub create_super_user: Option<String>,
 
     /// The logging verbosity
     #[arg(short, long)]
