--- backend/src/db/mod.rs
@@ -1,2 +1,0 @@
-pub mod migrator;
-pub mod users;
