+++ backend/src/handlers.rs
@@ -1,14 +1,11 @@
-use axum_login::AuthSession;
-
 use axum::{
-    Json,
     extract::State,
     http,
     response::{Html, IntoResponse},
 };
 use tokio::{fs, io};
 
-use crate::{auth, response_bodies, states::RootState};
+use crate::states::RootState;
 
 /// A response error
 #[derive(Debug)]
@@ -50,66 +47,4 @@ pub async fn get_index(State(state): State<RootState>) -> Result<impl IntoRespon
         .await
         .map(|content| Html(content).into_response())
         .map_err(|err| ErrorResponse::IoError(err))
-}
-
-pub mod backend {
-    use super::*;
-
-    pub async fn get_ping() -> impl IntoResponse {
-        "Pong".into_response()
-    }
-
-    pub async fn post_login(
-        mut auth_session: AuthSession<auth::Backend>,
-        Json(credentials): Json<auth::Credentials>,
-    ) -> impl IntoResponse {
-        match auth_session.authenticate(credentials).await {
-            Ok(Some(user)) => match auth_session.login(&user).await {
-                Ok(_) => (
-                    http::StatusCode::OK,
-                    Json(response_bodies::LoginResponse {
-                        username: user.username,
-                    }),
-                )
-                    .into_response(),
-                Err(err) => {
-                    (http::StatusCode::INTERNAL_SERVER_ERROR, format!("{}", err)).into_response()
-                }
-            },
-            Ok(None) => (http::StatusCode::UNAUTHORIZED, "Unauthorized").into_response(),
-            Err(err) => {
-                (http::StatusCode::INTERNAL_SERVER_ERROR, format!("{}", err)).into_response()
-            }
-        }
-    }
-
-    pub async fn post_logout(mut auth_session: AuthSession<auth::Backend>) -> impl IntoResponse {
-        match auth_session.user {
-            Some(_) => {
-                match auth_session.logout().await {
-                    Ok(_) => (http::StatusCode::OK, "OK").into_response(),
-                    Err(err) => (http::StatusCode::INTERNAL_SERVER_ERROR, format!("{}", err)).into_response()
-                }
-            },
-            None => (http::StatusCode::UNAUTHORIZED, "Unauthorized").into_response(),
-        }
-    }
-
-    pub async fn get_current_user(auth_session: AuthSession<auth::Backend>) -> impl IntoResponse {
-        match auth_session.user {
-            Some(user) => (
-                http::StatusCode::OK,
-                Json(response_bodies::MeResponse {
-                    username: user.username,
-                    admin: user.admin,
-                }),
-            )
-                .into_response(),
-            None => (http::StatusCode::UNAUTHORIZED, "Unauthorized").into_response(),
-        }
-    }
-
-    pub async fn get_404() -> impl IntoResponse {
-        (http::StatusCode::NOT_FOUND, "Not Found").into_response()
-    }
 }
