--- backend/src/response_bodies.rs
@@ -1,12 +1,0 @@
-use serde::Serialize;
-
-#[derive(Debug, Clone, Serialize)]
-pub struct LoginResponse {
-    pub username: String,
-}
-
-#[derive(Debug, Clone, Serialize)]
-pub struct MeResponse {
-    pub username: String,
-    pub admin: bool,
-}
