+++ frontend/Cargo.toml
@@ -4,15 +4,8 @@ version = "0.0.0"
 edition = "2024"
 
 [dependencies]
-gloo-net = "0.6.0"
-serde = "1.0.219"
-serde_json = "1.0.140"
-urlencoding = "2.1.3"
-wasm-bindgen-futures = "0.4.50"
 wasm-logger = "0.2.0"
-web-sys = { version = "0.3.77", features = ["HtmlInputElement", "FormData", "HtmlFormElement"] }
 yew = { version = "0.21.0", features = ["csr"] }
 yew-autoprops = "0.4.1"
-yew-hooks = "0.3.3"
 yew-router = "0.18.0"
-yewdux = "0.11.0"
+
