--- frontend/src/app/queries.rs
@@ -1,9 +1,0 @@
-use serde::{Deserialize, Serialize};
-
-use super::Route;
-
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub(super) struct LoginQuery {
-    #[serde(default)]
-    pub next: Option<Route>,
-}
