+++ frontend/src/main.rs
@@ -1,7 +1,6 @@
 use app::App;
 
 mod app;
-mod net;
 
 fn main() {
     wasm_logger::init(wasm_logger::Config::default());
