--- frontend/src/net/bodies.rs
@@ -1,7 +1,0 @@
-use serde::Serialize;
-
-#[derive(Debug, Clone, Serialize)]
-pub struct LoginBody {
-    pub username: String,
-    pub password: String,
-}
