--- frontend/src/net/mod.rs
@@ -1,2 +1,0 @@
-pub mod bodies;
-pub mod responses;
