--- frontend/src/net/responses.rs
@@ -1,7 +1,0 @@
-use serde::Deserialize;
-
-#[derive(Debug, Clone, Deserialize)]
-pub struct CurrentUserResponse {
-    pub username: String,
-    pub admin: bool,
-}
